requirements.txt
asgiref==3.10.0
bcrypt==5.0.0
blinker==1.9.0
cassandra-driver==3.29.3
certifi==2025.10.5
click==8.3.0
Django==5.2.7
dnspython==2.8.0
et_xmlfile==2.0.0
eventlet==0.40.3
Flask==3.1.2
flask-cors==6.0.1
geomet==1.1.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
neo4j==6.0.3
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pyaml==25.7.0
pymongo==4.15.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
clickhouse-connect>=0.6.1

